+++ src/main/java/org/opensearch/knn/index/mapper/KNNVectorFieldMapper.java
@@ -603,7 +603,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
             context,
             fieldType().getDimension(),
             fieldType().getSpaceType(),
-            getMethodComponentContext(fieldType().getKnnMethodContext())
+            getMethodComponentContext(fieldType().getKnnMethodContext()),
+            fieldType().getVectorDataType()
         );
     }
 
@@ -646,8 +647,13 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
         return fields;
     }
 
-    protected void parseCreateField(ParseContext context, int dimension, SpaceType spaceType, MethodComponentContext methodComponentContext)
-        throws IOException {
+    protected void parseCreateField(
+        ParseContext context,
+        int dimension,
+        SpaceType spaceType,
+        MethodComponentContext methodComponentContext,
+        VectorDataType vectorDataType
+    ) throws IOException {
 
         validateIfKNNPluginEnabled();
         validateIfCircuitBreakerIsNotTriggered();