+++ src/main/java/org/opensearch/knn/index/mapper/ModelFieldMapper.java
@@ -62,6 +62,12 @@ public class ModelFieldMapper extends KNNVectorFieldMapper {
             );
         }
 
-        parseCreateField(context, modelMetadata.getDimension(), modelMetadata.getSpaceType(), modelMetadata.getMethodComponentContext());
+        parseCreateField(
+            context,
+            modelMetadata.getDimension(),
+            modelMetadata.getSpaceType(),
+            modelMetadata.getMethodComponentContext(),
+            modelMetadata.getVectorDataType()
+        );
     }
 }