+++ src/main/java/org/opensearch/knn/plugin/transport/TrainingModelTransportAction.java
@@ -51,7 +51,8 @@ public class TrainingModelTransportAction extends HandledTransportAction<Trainin
             NativeMemoryLoadStrategy.TrainingLoadStrategy.getInstance(),
             clusterService,
             request.getMaximumVectorCount(),
-            request.getSearchSize()
+            request.getSearchSize(),
+            request.getVectorDataType()
         );
 
         // Allocation representing size model will occupy in memory during training
@@ -68,7 +69,8 @@ public class TrainingModelTransportAction extends HandledTransportAction<Trainin
             modelAnonymousEntryContext,
             request.getDimension(),
             request.getDescription(),
-            clusterService.localNode().getEphemeralId()
+            clusterService.localNode().getEphemeralId(),
+            request.getVectorDataType()
         );
 
         KNNCounter.TRAINING_REQUESTS.increment();