+++ src/main/java/org/opensearch/knn/training/TrainingJob.java
@@ -16,7 +16,9 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.common.UUIDs;
 import org.opensearch.knn.common.KNNConstants;
+import org.opensearch.knn.index.IndexUtil;
 import org.opensearch.knn.index.KNNSettings;
+import org.opensearch.knn.index.VectorDataType;
 import org.opensearch.knn.jni.JNIService;
 import org.opensearch.knn.index.KNNMethodContext;
 import org.opensearch.knn.index.memory.NativeMemoryAllocation;
@@ -31,6 +33,8 @@ import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.util.Map;
 import java.util.Objects;
+
+import static org.opensearch.knn.index.util.Faiss.FAISS_BINARY_INDEX_DESCRIPTION_PREFIX;
 
 /**
  * Encapsulates all information required to generate and train a model.
@@ -66,7 +70,8 @@ public class TrainingJob implements Runnable {
         NativeMemoryEntryContext.AnonymousEntryContext modelAnonymousEntryContext,
         int dimension,
         String description,
-        String nodeAssignment
+        String nodeAssignment,
+        VectorDataType vectorDataType
     ) {
         // Generate random base64 string if one is not provided
         this.modelId = StringUtils.isNotBlank(modelId) ? modelId : UUIDs.randomBase64UUID();
@@ -84,7 +89,8 @@ public class TrainingJob implements Runnable {
                 description,
                 "",
                 nodeAssignment,
-                knnMethodContext.getMethodComponentContext()
+                knnMethodContext.getMethodComponentContext(),
+                vectorDataType
             ),
             null,
             this.modelId
@@ -181,6 +187,15 @@ public class TrainingJob implements Runnable {
                 KNNConstants.INDEX_THREAD_QTY,
                 KNNSettings.state().getSettingValue(KNNSettings.KNN_ALGO_PARAM_INDEX_THREAD_QTY)
             );
+
+            if (VectorDataType.BINARY == model.getModelMetadata().getVectorDataType()) {
+                trainParameters.put(
+                    KNNConstants.INDEX_DESCRIPTION_PARAMETER,
+                    FAISS_BINARY_INDEX_DESCRIPTION_PREFIX + trainParameters.get(KNNConstants.INDEX_DESCRIPTION_PARAMETER).toString()
+                );
+            }
+
+            IndexUtil.updateVectorDataTypeToParameters(trainParameters, model.getModelMetadata().getVectorDataType());
 
             byte[] modelBlob = JNIService.trainIndex(
                 trainParameters,